UCT-714: Refactored calling deprecated method inspection
