+++ src/com/magento/idea/magento2uct/inspections/php/CallMethodInspection.java
@@ -1,0 +1,69 @@
+/*
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+
+package com.magento.idea.magento2uct.inspections.php;
+
+import com.intellij.codeInspection.ProblemsHolder;
+import com.intellij.openapi.project.Project;
+import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiElementVisitor;
+import com.jetbrains.php.lang.inspections.PhpInspection;
+import com.jetbrains.php.lang.psi.elements.Method;
+import com.jetbrains.php.lang.psi.elements.MethodReference;
+import com.jetbrains.php.lang.psi.visitors.PhpElementVisitor;
+import com.magento.idea.magento2uct.packages.IssueSeverityLevel;
+import com.magento.idea.magento2uct.settings.UctSettingsService;
+import org.jetbrains.annotations.NotNull;
+
+public abstract class CallMethodInspection extends PhpInspection {
+
+    @Override
+    public @NotNull PsiElementVisitor buildVisitor(
+            final @NotNull ProblemsHolder problemsHolder,
+            final boolean isOnTheFly
+    ) {
+        return new PhpElementVisitor() {
+
+            @Override
+            public void visitPhpMethodReference(final MethodReference reference) {
+                final Project project = reference.getProject();
+                final UctSettingsService settings = UctSettingsService.getInstance(project);
+
+                if (!settings.isEnabled()
+                        || !settings.isIssueLevelSatisfiable(getSeverityLevel())) {
+                    return;
+                }
+                final PsiElement resolvedElement = reference.resolve();
+
+                if (!(resolvedElement instanceof Method)) {
+                    return;
+                }
+                execute(project, problemsHolder, reference, ((Method) resolvedElement));
+            }
+        };
+    }
+
+    /**
+     * Implement this method to specify inspection logic.
+     *
+     * @param project Project
+     * @param problemsHolder ProblemsHolder
+     * @param methodReference MethodReference
+     * @param method Method
+     */
+    protected abstract void execute(
+            final Project project,
+            final @NotNull ProblemsHolder problemsHolder,
+            final MethodReference methodReference,
+            final Method method
+    );
+
+    /**
+     * Implement this method to specify issue severity level for target inspection.
+     *
+     * @return IssueSeverityLevel
+     */
+    protected abstract IssueSeverityLevel getSeverityLevel();
+}
