+++ src/com/magento/idea/magento2uct/inspections/php/deprecation/CallingDeprecatedMethod.java
@@ -8,59 +8,44 @@ package com.magento.idea.magento2uct.inspections.php.deprecation;
 import com.intellij.codeInspection.ProblemHighlightType;
 import com.intellij.codeInspection.ProblemsHolder;
 import com.intellij.openapi.project.Project;
-import com.intellij.psi.PsiElement;
-import com.intellij.psi.PsiElementVisitor;
-import com.jetbrains.php.lang.inspections.PhpInspection;
 import com.jetbrains.php.lang.psi.elements.Method;
 import com.jetbrains.php.lang.psi.elements.MethodReference;
-import com.jetbrains.php.lang.psi.visitors.PhpElementVisitor;
 import com.magento.idea.magento2uct.inspections.UctProblemsHolder;
+import com.magento.idea.magento2uct.inspections.php.CallMethodInspection;
+import com.magento.idea.magento2uct.packages.IssueSeverityLevel;
 import com.magento.idea.magento2uct.packages.SupportedIssue;
-import com.magento.idea.magento2uct.settings.UctSettingsService;
 import com.magento.idea.magento2uct.versioning.VersionStateManager;
 import org.jetbrains.annotations.NotNull;
 
-public class CallingDeprecatedMethod extends PhpInspection {
+public class CallingDeprecatedMethod extends CallMethodInspection {
 
     @Override
-    public @NotNull PsiElementVisitor buildVisitor(
+    protected void execute(
+            final Project project,
             final @NotNull ProblemsHolder problemsHolder,
-            final boolean isOnTheFly
+            final MethodReference methodReference,
+            final Method method
     ) {
-        return new PhpElementVisitor() {
+        final String type = method.getFQN();
 
-            @Override
-            public void visitPhpMethodReference(final MethodReference reference) {
-                final Project project = reference.getProject();
-                final UctSettingsService settings = UctSettingsService.getInstance(project);
-
-                if (!settings.isEnabled() || !settings.isIssueLevelSatisfiable(
-                        SupportedIssue.CALLING_DEPRECATED_METHOD.getLevel())
-                ) {
-                    return;
-                }
-                final PsiElement resolvedElement = reference.resolve();
-
-                if (!(resolvedElement instanceof Method)) {
-                    return;
-                }
-                final String type = ((Method) resolvedElement).getFQN();
-
-                if (VersionStateManager.getInstance(project).isDeprecated(type)) {
-                    if (problemsHolder instanceof UctProblemsHolder) {
-                        ((UctProblemsHolder) problemsHolder).setReservedErrorCode(
-                                SupportedIssue.CALLING_DEPRECATED_METHOD.getCode()
-                        );
-                    }
-                    problemsHolder.registerProblem(
-                            reference,
-                            SupportedIssue.CALLING_DEPRECATED_METHOD.getMessage(
-                                    type.replace(".", "::")
-                            ),
-                            ProblemHighlightType.LIKE_DEPRECATED
-                    );
-                }
+        if (VersionStateManager.getInstance(project).isDeprecated(type)) {
+            if (problemsHolder instanceof UctProblemsHolder) {
+                ((UctProblemsHolder) problemsHolder).setReservedErrorCode(
+                        SupportedIssue.CALLING_DEPRECATED_METHOD.getCode()
+                );
             }
-        };
+            problemsHolder.registerProblem(
+                    methodReference,
+                    SupportedIssue.CALLING_DEPRECATED_METHOD.getMessage(
+                            type.replace(".", "::")
+                    ),
+                    ProblemHighlightType.LIKE_DEPRECATED
+            );
+        }
+    }
+
+    @Override
+    protected IssueSeverityLevel getSeverityLevel() {
+        return SupportedIssue.CALLING_DEPRECATED_METHOD.getLevel();
     }
 }
